Support for the first game state - RUNNING
