+++ pycraft/main.py
@@ -23,8 +23,8 @@ def main():
     )
     # Hide the mouse cursor and prevent the mouse from leaving the window.
     window.set_exclusive_mouse(config_data["window"]["exclusive_mouse"])
-    world = World()
-    window.set_world(world)
-    player = Player(config_data["world"])
-    window.set_player(player)
+    # world = World()
+    # window.set_world(world)
+    # player = Player(config_data["world"])
+    # window.set_player(player)
     pyglet.app.run()
